A number of PR changes have been implemented.
-.c and .h files split five ways from main_functions
-defined macros for modes (for consistency, 1 is quiet, 2 verbose, 3 interactive)
-booleans for fileexists, quit
-formatting for header files
-underline_misspelled/underline_correct_spelling erased
-parse_string changed to void
-testing expansion started

Questions remain on my part whether char** array is always the wrong data type to use for a variety of functions.
