+++ src/shellstrings.c
@@ -80,11 +80,6 @@ void shell_help() {
 }
 
 /* See shellstrings.h */
-void shell_outro() {
-    printf("Thank you for using Spellcheck. Bye!\n");
-}
-
-/* See shellstrings.h */
 void shell_verbose_chart(int lnum, char* misspelled, char** suggestions) {
     printf("%d\t\t\t", lnum); // print line number
     int ntab = 3 - (strlen(misspelled) / 8); // number of tabs
