+++ tests/test_main_functions_interactive.c
@@ -1,44 +1,18 @@
 #include <criterion/criterion.h>
 #include <stdlib.h>
 #include <stdbool.h>
-#include "main_functions.h"
+#include "main_functions_edit.h"
 #include <string.h>
 
-/*
- ***** underline_misspelled tests *****
+/* Note on testing for main_functions_*.c files:
+ * Note that there are five parts to main_functions.c: saving, editing, interactive, batch, main.
+ * The functions that do not require command line inputs have been tested in the test_main_functions files.
+ * The functions that do require command line inputs are currently being researched by Deniz Turkcapar,
+ * but currently use a simpler testing version, via command lines accepted in travis, to ensure that
+ * each function responsible for printing a page prints a correct output.
+ * For this reason, test_main_functions_interactive.c does not exist, given that the edit_interactive function, and
+ * the wrapper interactive_mode function, cannot operate without command line inputs.
  */
-void check_underline_misspelled(char* wrong, char* underline, char* expected) {
-	underline_misspelled(wrong, underline);
-
-	int result = strcmp(underline, expected);
-
-    cr_assert_eq(result, 0, "underline_misspelled_mispelled failed");
-}
-
-Test(main_functions, underline_misspelled) {
-	char *wrong = "bool";
-	char *underline = (char *)malloc(50);
-	strcpy(underline, "");
-	char* expected = "^^^^ ";
-    check_underline_misspelled(wrong, underline, expected);
-}
-
-Test(main_functions, underline_misspelled2) {
-	char *wrong = "hello";
-	char *underline = (char *)malloc(50);
-	strcpy(underline, "    ");
-	char* expected = "    ^^^^^ ";
-    check_underline_misspelled(wrong, underline, expected);
-}
-
-
-Test(main_functions, underline_misspelled3) {
-	char *wrong = "wrng";
-	char *underline = (char *)malloc(50);
-	strcpy(underline, "^^^^ ^^ ");
-	char* expected = "^^^^ ^^ ^^^^ ";
-    check_underline_misspelled(wrong, underline, expected);
-}
 
 
 /*
@@ -59,14 +33,12 @@ void check_add_to_misspelled(char *word, char** misspelled) {
 
 }
 
-Test(main_functions, check_add_to_misspelled) {
+Test(main_functions_edit, check_add_to_misspelled) {
 	char *badword = "wrod";
 	char* badwords[3] = {NULL, NULL, NULL};
 	check_add_to_misspelled(badword, badwords);
 
 }
-
-
 
 /*
  ***** correct_line tests *****
@@ -80,7 +52,7 @@ void test_correct_line(char* line, char* old_word, char* new_word, char* expecte
     cr_assert_eq(result, 0, "correct_line test failed");
 }
 
-Test(main_functions, correct_line1) {
+Test(main_functions_edit, correct_line1) {
 	char *line_code = "I don't know hw to spell";
 	char *line = (char *)malloc(strlen(line_code)+1);
 	strcpy(line, line_code);
@@ -95,7 +67,7 @@ Test(main_functions, correct_line1) {
     test_correct_line(line, old_word, new_word, expected);
 }
 
-Test(main_functions, correct_line2) {
+Test(main_functions_edit, correct_line2) {
 	char *line_code = "I dont know how to spell";
 	char *line = (char *)malloc(strlen(line_code)+1);
 	strcpy(line, line_code);
@@ -110,7 +82,7 @@ Test(main_functions, correct_line2) {
     test_correct_line(line, old_word, new_word, expected);
 }
 
-Test(main_functions, correct_line3) {
+Test(main_functions_edit, correct_line3) {
 	char *line_code = "spl";
 	char *line = (char *)malloc(strlen(line_code)+1);
 	strcpy(line, line_code);
@@ -137,7 +109,7 @@ void check_underline_misspelled_sentence(char** badwords, char* sentence, char* 
     cr_assert_eq(result, 0, "underline_misspelled test failed");
 }
 
-Test(main_functions, underline_misspelled_sentence) {
+Test(main_functions_edit, underline_misspelled_sentence) {
 	char **badwords = (char **)malloc((sizeof(char*)*3));
 	char *line = "bd splling is bad";
 	char *underline = (char *)malloc(sizeof(char) * (strlen(line) + 1));
@@ -150,7 +122,7 @@ Test(main_functions, underline_misspelled_sentence) {
 }
 
 
-Test(main_functions, underline_misspelled_sentence2) {
+Test(main_functions_edit, underline_misspelled_sentence2) {
 	char *line = "bd splling is bd";
 	char **badwords = (char **)malloc((sizeof(char*)*4));
 	char *underline = (char *)malloc(sizeof(char) * (strlen(line) + 1));
@@ -176,20 +148,20 @@ void check_is_in_punct_array(char letter, int expected) {
 }
 
 
-Test(main_functions, is_in_punct_array) {
+Test(main_functions_edit, is_in_punct_array) {
     check_is_in_punct_array('"', EXIT_SUCCESS);
 }
 
-Test(main_functions, is_in_punct_array3) {
+Test(main_functions_edit, is_in_punct_array3) {
     check_is_in_punct_array('?', EXIT_SUCCESS);
 }
 
-Test(main_functions, is_in_punct_array4) {
+Test(main_functions_edit, is_in_punct_array4) {
     check_is_in_punct_array('0', EXIT_FAILURE);
 }
 
 
-Test(main_functions, is_in_punct_array2) {
+Test(main_functions_edit, is_in_punct_array2) {
     check_is_in_punct_array('a', EXIT_FAILURE);
 }
 
@@ -206,19 +178,19 @@ void check_remove_prefix_punctuation(char* word, char* expected) {
     cr_assert_eq(result, 0, "remove_prefix_punctuation test failed");
 }
 
-Test(main_functions, remove_prefix_punctuation) {
+Test(main_functions_edit, remove_prefix_punctuation) {
 	char buffer[8] = "...words";
     char* expected = "words"; 
     check_remove_prefix_punctuation(buffer, expected);
 }
 
-Test(main_functions, remove_prefix_punctuation2) {
+Test(main_functions_edit, remove_prefix_punctuation2) {
 	char buffer[7] = "..words";
     char* expected = "words"; 
     check_remove_prefix_punctuation(buffer, expected);
 }
 
-Test(main_functions, remove_prefix_punctuation3) {
+Test(main_functions_edit, remove_prefix_punctuation3) {
 	char buffer[9] = "??words?";
     char* expected = "words?"; 
     check_remove_prefix_punctuation(buffer, expected);
@@ -236,7 +208,7 @@ void check_remove_trailing_punctuation(char* word, char* expected) {
     cr_assert_eq(result, 0, "remove_trailing_punctuation test failed");
 }
 
-Test(main_functions, remove_trailing_punctuation1) {
+Test(main_functions_edit, remove_trailing_punctuation1) {
     char* sample = "...words...";
     char *buffer = (char *) malloc(sizeof(char) * (strlen(sample) + 1));
 	strcpy(buffer, sample);
@@ -244,19 +216,19 @@ Test(main_functions, remove_trailing_punctuation1) {
     check_remove_trailing_punctuation(buffer, expected);
 }
 
-Test(main_functions, remove_trailing_punctuation2) {
+Test(main_functions_edit, remove_trailing_punctuation2) {
 	char buffer[6] = "words.";
     char* expected = "words"; 
     check_remove_trailing_punctuation(buffer, expected);
 }
 
-Test(main_functions, remove_trailing_punctuation3) {
+Test(main_functions_edit, remove_trailing_punctuation3) {
 	char buffer[8] = "??words?";
     char* expected = "??words"; 
     check_remove_trailing_punctuation(buffer, expected);
 }
 
-Test(main_functions, remove_trailing_punctuation4) {
+Test(main_functions_edit, remove_trailing_punctuation4) {
     char* sample = "words.!?!?!.";
     char *buffer = (char *) malloc(sizeof(char) * (strlen(sample) + 1));
 	strcpy(buffer, sample);
