+++ tests/test_short.txt
@@ -1,2 +1,2 @@
-Eye have, a splling chequer 
+Eye have, a splling chequer
 It cme with thy m'y PC.
